pyproject.toml
[project]
name = "acadextract"
version = "1.0.0"
description = "Autonomous Academic Result Extraction and Student Profiling from Email Streams"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "AcadExtract Team"}
]

dependencies = [
    # Web framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.23",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.2.4",

    # Task queue
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",

    # Email
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "imapclient>=2.3.1",

    # PDF parsing
    "pdfplumber>=0.10.0",
    "camelot-py[cv]>=0.11.0",
    "pymupdf>=1.23.0",
    "pdf2image>=1.16.0",
    "tabula-py>=2.8.0",

    # OCR
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.5.0",
    "pytesseract>=0.3.10",

    # Excel
    "openpyxl>=3.1.0",
    "pandas>=2.1.0",

    # ML / NLP
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",

    # LLM
    "openai>=1.3.0",
    "google-generativeai>=0.3.0",

    # Storage
    "boto3>=1.29.0",
    "minio>=7.2.0",

    # Utilities
    "pyyaml>=6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "tenacity>=8.2.0",
    "simhash>=2.1.2",
    "datasketch>=1.6.0",
    "sqlglot>=19.0.0",
    "jellyfish>=1.0.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",

    # Observability
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-exporter-otlp>=1.21.0",
    "sentry-sdk[fastapi]>=1.38.0",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
    "factory-boy>=3.3.0",
    "faker>=20.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "COM", "C4", "DTZ", "T20", "ISC", "ICN", "PIE", "PT", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "ERA"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
